OR-130:Fixed the issue of routing from payment page
* Now user will be routed from payment page to order confirmation page
* now confirm call will happen on the mount of confirmation page
* removed status call from order detail page as it is not needed there
