+++ pages/paymentMode.tsx
@@ -1,14 +1,27 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Box, Text } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { AppHeader } from "../components/appHeader/AppHeader";
 import Button from "../components/button/Button";
 import CardWithCheckBox from "../components/card/Card";
 import { useLanguage } from "../hooks/useLanguage";
+import { useSelector } from "react-redux";
+import { getSubTotalAndDeliveryCharges } from "../utilities/checkout-utils";
 
 function PaymentMode() {
+  const [totalPrice, setTotalPrice] = useState(0);
   const { t } = useLanguage();
   const router = useRouter();
+
+  const initData = useSelector((state: any) => state.initResponse.initResponse);
+
+  useEffect(() => {
+    if (initData) {
+      const { subTotal, totalDeliveryCharge } =
+        getSubTotalAndDeliveryCharges(initData);
+      setTotalPrice(subTotal + totalDeliveryCharge);
+    }
+  }, []);
 
   return (
     <>
@@ -19,11 +32,11 @@ function PaymentMode() {
       </Box>
 
       <Button
-        buttonText={t.proceedToPay}
+        buttonText={`${t.proceedToPay} Rs.${totalPrice}`}
         background={"rgba(var(--color-primary))"}
         color={"rgba(var(--text-color))"}
         isDisabled={false}
-        handleOnClick={() => router.push("/orderDetails")}
+        handleOnClick={() => router.push("/orderConfirmation")}
       />
     </>
   );
